Conflicts:
	src/plugins/qmldesigner/settingspage.ui
	src/plugins/qt4projectmanager/qt-maemo/maemopackagecontents.cpp
	src/plugins/qt4projectmanager/qt-maemo/maemorunconfiguration.cpp
	src/plugins/qt4projectmanager/qt-s60/s60createpackagestep.cpp

+++ src/plugins/qt4projectmanager/qt-maemo/maemodeviceconfigurations.h
@@ -96,7 +96,7 @@ public:
     QList<MaemoDeviceConfig> devConfigs() const { return m_devConfigs; }
     void setDevConfigs(const QList<MaemoDeviceConfig> &devConfigs);
     MaemoDeviceConfig find(const QString &name) const;
-    MaemoDeviceConfig find(int id) const;
+    MaemoDeviceConfig find(quint64 id) const;
 
 signals:
     void updated();
